requirements.txt
git+https://www.github.com/keras-team/keras-contrib.git
CNLTK-Beta==1.0.1
joblib
tensorflow
keras
plot-keras-history
tensorflow_addons
keras-self-attention
attention
nltk
git+https://www.github.com/keras-team/keras-contrib.git
flask
gunicorn